simple new-request componets are now TS
