+++ src/client/components/main/WebRTC-composer/WebRTCComposer.tsx
@@ -10,8 +10,8 @@ import historyController from '../../../controllers/historyController';
  */
 import WebRTCSessionEntryForm from './WebRTCSessionEntryForm.jsx';
 import WebRTCServerEntryForm from './WebRTCServerEntryForm.jsx';
-import NewRequestButton from '../new-request/NewRequestButton.jsx';
-import TestEntryForm from '../new-request/TestEntryForm.jsx';
+import NewRequestButton from '../new-request/NewRequestButton';
+import TestEntryForm from '../new-request/TestEntryForm';
 // Import MUI components
 import { Box } from '@mui/material';
