--- src/client/components/main/new-request/NewRequestButton.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const NewRequestButton = ({ onClick }) => (
-  <button
-    className="button is-normal is-primary-100 add-request-button is-vertical-align-center is-justify-content-center no-border-please"
-    onClick={onClick}
-    type="button"
-    style={{margin: '10px'}}
-  >
-    Add to Workspace
-  </button>
-);
-
-export default NewRequestButton;
